// src/front/src/api/answerrule/models/answeringRuleAction.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX AnswerRule API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export type AnsweringRuleAction =
	(typeof AnsweringRuleAction)[keyof typeof AnsweringRuleAction];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AnsweringRuleAction = {
	Voicemail: "Voicemail",
	Redirect: "Redirect",
	RedirectToGroup: "RedirectToGroup",
	DisconnectWithVoicemessage: "DisconnectWithVoicemessage",
} as const;