// src/front/src/api/answerrule/models/answeringRuleResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX AnswerRule API (via Gateway)
 * OpenAPI spec version: 1.0
 */
import type { AnsweringRuleAction } from "./answeringRuleAction";

export interface AnsweringRuleResponse {
	/** @nullable */
	gid?: string | null;
	/** @nullable */
	name?: string | null;
	/** @nullable */
	description?: string | null;
	priority?: number;
	isEnabled?: boolean;
	actionType?: AnsweringRuleAction;
	/** @nullable */
	actionTarget?: string | null;
	sendEmailNotification?: boolean;
	timeSlotsCount?: number;
	createdAt?: string;
}